if ($downloadInfo['status'] == Common::DOWNLOAD_DELETION_ENABLE) {                 $this->downloadFilesService->deleteInfo([                     'id' => $downloadInfo['id']                 ]);             } else {                 dd('Can download the torrent only after hls cut and upload');             }

+++ app/Console/Commands/TorrentDownloaderCommand.php
@@ -61,12 +61,16 @@ class TorrentDownloaderCommand extends Command
      */
     public function handle()
     {
-        $downloadInfo = $this->downloadFilesService->getInfo([['status', '=', Common::DOWNLOAD_DELETION_ENABLE]], ['*'], ['id', 'DESC']);
+        $downloadInfo = $this->downloadFilesService->getInfo([], ['*'], ['id', 'DESC']);
         
         if ($downloadInfo) {
-            $this->downloadFilesService->deleteInfo([
-                'id' => $downloadInfo['id']
-            ]);
+            if ($downloadInfo['status'] == Common::DOWNLOAD_DELETION_ENABLE) {
+                $this->downloadFilesService->deleteInfo([
+                    'id' => $downloadInfo['id']
+                ]);
+            } else {
+                dd('Can download the torrent only after hls cut and upload');
+            }
         }
         
         $originalSource = $this->pickUpOneItem();
